add union, minus, and filter
